# live-api-documentation.yaml
openapi: 3.0.0
info:
  title: Live API Gateway
  version: "2.0.0"
  description: >
    API desarrollada con Express en Firebase Functions que integra funciones de
    verificación de email, feed de stories, eventos, seguimiento, likes y más.

    **VERSIÓN 2.0 - MIGRADA A NUEVA ESTRUCTURA**
    - Todas las APIs han sido migradas a la nueva estructura /races/{raceId}/events/{eventId}/
    - Solo incluye APIs activamente implementadas y migradas
    - APIs obsoletas/duplicadas han sido removidas

    **OPTIMIZACIONES ULTRA-RÁPIDAS**
    - Feed Extended API: 99% más rápido con Collection Group Queries
    - Rendimiento: Historia específica ~50ms, Feed completo ~150ms
    - Paginación eficiente optimizada para aplicaciones móviles
    - Consultas paralelas y enriquecimiento optimizado
  contact:
    name: Live Copernico Team
    email: support@copernico.com
servers:
  - url: https://us-central1-live-copernico.cloudfunctions.net/liveApiGateway
    description: Servidor de producción Firebase
  - url: http://127.0.0.1:5001/live-copernico/us-central1/liveApiGateway
    description: Servidor de desarrollo local
paths:
...
  /api/generateUploadUrl:
    post:
      tags:
        - Upload
      summary: Generar URL prefirmada para subida a Firebase Storage
      description: >
        Genera una URL prefirmada para subir archivos directamente a Firebase Storage.
        MIGRADO a Firebase Storage y nueva estructura.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                raceId:
                  type: string
                  description: Identificador de la carrera (NUEVO - requerido)
                  example: "race123"
                eventId:
                  type: string
                  description: Identificador del evento
                  example: "event456"
                participantId:
                  type: string
                  description: Identificador del participante
                  example: "participant789"
                fileName:
                  type: string
                  description: Nombre del archivo
                  example: "video.mp4"
                contentType:
                  type: string
                  description: Tipo de contenido del archivo
                  example: "video/mp4"
              required:
                - raceId
                - eventId
                - participantId
                - fileName
                - contentType
      responses:
        '200':
          description: URL prefirmada generada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                    description: URL prefirmada para subida
                  filePath:
                    type: string
                    description: Ruta del archivo en Storage
                  expiresAt:
                    type: string
                    format: date-time
                    description: Fecha de expiración de la URL
        '400':
          description: Parámetros faltantes
        '500':
          description: Error interno del servidor
  /api/uploadToFirebase:
    post:
      tags:
        - Upload
      summary: Subir archivo directamente a Firebase Storage
      description: >
        Sube un archivo directamente a Firebase Storage con metadata completa.
        MIGRADO a Firebase Storage y nueva estructura.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo a subir
                raceId:
                  type: string
                  description: Identificador de la carrera (NUEVO - requerido)
                  example: "race123"
                eventId:
                  type: string
                  description: Identificador del evento
                  example: "event456"
                participantId:
                  type: string
                  description: Identificador del participante
                  example: "participant789"
                description:
                  type: string
                  description: Descripción del archivo (opcional)
                  example: "Video del participante"
              required:
                - file
                - raceId
                - eventId
                - participantId
      responses:
        '200':
          description: Archivo subido exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Archivo subido exitosamente"
                  fileUrl:
                    type: string
                    example: "https://storage.googleapis.com/..."
                  fileName:
                    type: string
                    example: "abc123-video.mp4"
                  mediaType:
                    type: string
                    enum: [image, video, unknown]
        '400':
          description: Parámetros faltantes o archivo no válido
        '500':
          description: Error interno del servidor
  /api/confirmUpload:
    post:
      tags:
        - Upload
      summary: Confirmar subida y registrar en Firestore
      description: >
        Confirma la subida de un archivo y registra la metadata en Firestore.
        MIGRADO a nueva estructura.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                raceId:
                  type: string
                  description: Identificador de la carrera (NUEVO - requerido)
                  example: "race123"
                eventId:
                  type: string
                  description: Identificador del evento
                  example: "event456"
                participantId:
                  type: string
                  description: Identificador del participante
                  example: "participant789"
                fileName:
                  type: string
                  description: Nombre del archivo subido
                  example: "abc123-video.mp4"
                fileUrl:
                  type: string
                  description: URL del archivo en Storage
                  example: "https://storage.googleapis.com/..."
                description:
                  type: string
                  description: Descripción del archivo (opcional)
                  example: "Video del participante"
                originType:
                  type: string
                  description: Origen del archivo
                  example: "manual-upload"
              required:
                - raceId
                - eventId
                - participantId
                - fileName
                - fileUrl
      responses:
        '200':
          description: Upload confirmado y registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Upload confirmado exitosamente"
                  documentId:
                    type: string
                    example: "doc123abc"
                  storyData:
                    type: object
                    description: Datos de la historia registrada
        '400':
          description: Parámetros faltantes
        '500':
          description: Error interno del servidor
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para autenticación
...